+++ geom3d/paths/__init__.py
@@ -1,6 +1,7 @@
 from .path import Path
 from .polygon import cover_polygon2d_with_path
-from .nonintersecting import make_nonintersecting, MakeNonintersectingPaths
+from .nonintersecting import make_nonintersecting, MakeNonintersectingPaths, are_mutually_nonintersecting
 
 __all__ = ['Path', 'cover_polygon2d_with_path',
-           'make_nonintersecting', 'MakeNonintersectingPaths']
+           'make_nonintersecting', 'MakeNonintersectingPaths',
+           'are_mutually_nonintersecting']